vehicle_drift/requirements.txt
numpy>=1.21.0
torch>=1.9.0
mujoco>=2.3.0
# pychrono>=7.0.0
pyyaml>=6.0
gym>=0.21.0
matplotlib>=3.5.0
tensorboard>=2.8.0 